pyproject.toml
[tool.poetry]
name = "et.al-backend"
version = "0.1.0"
description = "Backed for the Research Assistant App"
authors = ["Geet Sethi <geetsethi05@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["all"], version = "^0.111.0" }
python-dotenv = "^1.0.1"
uvicorn = "^0.30.1"
pydantic = { extras = ["email"], version = "^2.7.4" }
sqlmodel = "^0.0.19"
psycopg2-binary = "^2.9.9"
argon2-cffi = "^23.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
